--- src/main/java/shapes/old/DBLayer/MongoDBRepository.java
@@ -1,36 +1,0 @@
-package shapes.old.DBLayer;
-
-public class MongoDBRepository {
-
-/*
-
-    public double calculateArea(String json) {
-        Shape shape = Converter.jsonToShapes(json).get(0);
-        return shape.calculateArea();
-    }
-
-    public Shape scaleShape(Map<String, String> params) {
-        String json = params.get("json");
-        double scale = Double.parseDouble(params.get("scale"));
-        Shape shape = Converter.jsonToShapes(json).get(0);
-        shape.changeSize(scale);
-        return shape;
-    }
-
-    public Shape moveShape(Map<String, String> params) {
-        String json = params.get("json");
-        double x = Double.parseDouble(params.get("x"));
-        double y = Double.parseDouble(params.get("y"));
-        Shape shape = Converter.jsonToShapes(json).get(0);
-        shape.move(x,y);
-        return shape;
-    }
-
-    public Shape rollShape(Map<String, String> params) {
-        String json = params.get("json");
-        double angle = Double.parseDouble(params.get("angle"));
-        Shape shape = Converter.jsonToShapes(json).get(0);
-        shape.roll(angle);
-        return shape;
-    }*/
-}
