--- src/main/java/shapes/old/ZKController.java
@@ -1,25 +1,0 @@
-package shapes.old;
-
-public class ZKController {
-/*
-    @RequestMapping(value = "calcShapeArea", method = RequestMethod.POST)
-    public double calculateArea(@RequestBody byte[] bytes) {
-        String json = new String(bytes, StandardCharsets.UTF_8);
-        return mongoDBRepository.calculateArea(json);
-    }
-
-    @RequestMapping(value = "scaleShape", method = RequestMethod.POST)
-    public String scaleShape(@RequestBody byte[] bytes) {
-        return Converter.shapeToJSON(mongoDBRepository.scaleShape(Converter.bytesToMap(bytes)));
-    }
-
-    @RequestMapping(value = "moveShape", method = RequestMethod.POST)
-    public String moveShape(@RequestBody byte[] bytes) {
-        return Converter.shapeToJSON(mongoDBRepository.moveShape(Converter.bytesToMap(bytes)));
-    }
-
-    @RequestMapping(value = "rollShape", method = RequestMethod.POST)
-    public String rollShape(@RequestBody byte[] bytes) {
-        return Converter.shapeToJSON(mongoDBRepository.rollShape(Converter.bytesToMap(bytes)));
-    }*/
-}
