+++ src/main/java/shapes/old/utils/Utils.java
@@ -1,10 +1,9 @@
 package shapes.old.utils;
 
-import shapes.old.Shapes.Point;
+import shapes.models.Point;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.ArrayList;
 import java.util.List;
 
 public class Utils {
@@ -16,26 +15,11 @@ public class Utils {
             centerX += point.getX();
             centerY += point.getY();
         }
-        return new Point(centerX / points.size(), centerY / points.size());
+        Point p = new Point();
+        p.setX(centerX / points.size());
+        p.setY(centerY / points.size());
+        return p;
     }
-
-    public static List<Point> toPoints(List<Double> params) {
-        List<Point> points = new ArrayList<>();
-        for (int i = 0; i < params.size() - 1; i = i + 2) {
-            Point point = new Point(params.get(i), params.get(i + 1));
-            points.add(point);
-        }
-        return points;
-    }
-
-    public static String getPointsAsString(List<Point> points) {
-        StringBuilder pointCoords = new StringBuilder();
-        for (Point point : points) {
-            pointCoords.append("(").append(point.getX()).append(", ").append(point.getY()).append(") ");
-        }
-        return pointCoords.toString().trim();
-    }
-
     /**
      * Возвращает округленное число типа double в формате #.##
      */
